Tweaks to pass CI tests

+++ src/providers/mbed_provider/utils.rs
@@ -1,8 +1,8 @@
 // Copyright 2019 Contributors to the Parsec project.
 // SPDX-License-Identifier: Apache-2.0
-use psa_crypto::types::key;
 use parsec_interface::operations::psa_key_attributes::Type;
 use parsec_interface::requests::{ResponseStatus, Result};
+use psa_crypto::types::key;
 
 macro_rules! bits_to_bytes {
     ($size:expr) => {
@@ -14,8 +14,8 @@ macro_rules! bits_to_bytes {
 /// Implementing `PSA_ASYMMETRIC_SIGN_OUTPUT_SIZE` as defined in `crypto_sizes.h` (Mbed Crypto).
 pub fn psa_asymmetric_sign_output_size(key_attrs: &key::Attributes) -> Result<usize> {
     match key_attrs.key_type {
-        Type::RsaKeyPair => Ok(usize::from(bits_to_bytes!(key_attrs.bits))),
-        Type::EccKeyPair{ .. } => Ok(usize::from(bits_to_bytes!(key_attrs.bits) * 2)),
+        Type::RsaKeyPair => Ok(bits_to_bytes!(key_attrs.bits)),
+        Type::EccKeyPair { .. } => Ok(bits_to_bytes!(key_attrs.bits) * 2),
         _ => Err(ResponseStatus::PsaErrorNotSupported),
     }
 }
@@ -30,9 +30,7 @@ pub fn psa_export_public_key_size(key_attrs: &key::Attributes) -> Result<usize> 
     };
 
     match key_attrs.key_type {
-        Type::RsaPublicKey | Type::RsaKeyPair => Ok(usize::from(
-            export_asn1_int_max_size!(key_attrs.bits) + 11,
-        )),
+        Type::RsaPublicKey | Type::RsaKeyPair => Ok(export_asn1_int_max_size!(key_attrs.bits) + 11),
         _ => Err(ResponseStatus::PsaErrorNotSupported),
     }
 }
